requirements.txt
# Core dependencies
PyYAML==6.0.1
requests==2.31.0

# Path expressions (will add later when needed)
# jsonpath-ng==1.6.1
# lxml==5.1.0

# Development dependencies (add when testing)
# pytest==7.4.3
# pytest-cov==4.1.0
